---ALPHA 3--- -Added public static Expand() to DB_RECORD to be able to define expanded/encrypted fields (after subclass declaration) without overriding the constructor
-DB_RECORD got some ORM support:
Added public static Relate($field,$class) to be able to directly access records referenced by their key via column name's overloaded property.
Setting properties marked as related to a non-object directly sets it's value to the given value, while setting it to a corresponding record (an instance of the class assigned to the column) sets it to this record's primary key.
Only subclasses of DB_RECORD can be set in relation mappings.

-Added public Copy() for cloning records. If the copied record has a primary key, it will be resetted to null.
-Added public static SGetPKField() to check a table's PK in static context (needed for internal checks)

-Added DB_RELATIONSTORE class (for internal usage with ORM functions)

-DUMMY now creates its optional log entry on construct instead of every overloaded function call